+++ rotkehlchen/tests/fixtures/rotkehlchen.py
@@ -1,18 +1,21 @@
 import base64
 from collections import namedtuple
+from contextlib import ExitStack
 from unittest.mock import patch
 
 import pytest
 
 import rotkehlchen.tests.utils.exchanges as exchange_tests
 from rotkehlchen.args import DEFAULT_MAX_LOG_BACKUP_FILES, DEFAULT_MAX_LOG_SIZE_IN_MB
-from rotkehlchen.db.settings import DBSettings
+from rotkehlchen.data_migrations.manager import DataMigrationManager
+from rotkehlchen.db.settings import LAST_DATA_MIGRATION, DBSettings
 from rotkehlchen.exchanges.manager import EXCHANGES_WITH_PASSPHRASE
 from rotkehlchen.history.price import PriceHistorian
 from rotkehlchen.premium.premium import Premium, PremiumCredentials
 from rotkehlchen.rotkehlchen import Rotkehlchen
 from rotkehlchen.tests.utils.api import create_api_server
 from rotkehlchen.tests.utils.database import (
+    _use_prepared_db,
     add_blockchain_accounts_to_db,
     add_manually_tracked_balances_to_test_db,
     add_settings_to_test_db,
@@ -65,6 +68,11 @@ def fixture_rotki_premium_credentials() -> PremiumCredentials:
     )
 
 
+@pytest.fixture(name='data_migration_version', scope='session')
+def fixture_data_migration_version() -> int:
+    return LAST_DATA_MIGRATION
+
+
 @pytest.fixture(name='cli_args')
 def fixture_cli_args(data_dir, ethrpc_endpoint):
     args = namedtuple('args', [
@@ -87,6 +95,12 @@ def fixture_cli_args(data_dir, ethrpc_endpoint):
     args.max_size_in_mb_all_logs = DEFAULT_MAX_LOG_SIZE_IN_MB
     args.max_logfiles_num = DEFAULT_MAX_LOG_BACKUP_FILES
     return args
+
+
+@pytest.fixture(name='perform_migrations_at_unlock')
+def fixture_perform_migrations_at_unlock():
+    """Perform data migrations as normal during user unlock"""
+    return True
 
 
 def initialize_mock_rotkehlchen_instance(
@@ -114,6 +128,10 @@ def initialize_mock_rotkehlchen_instance(
         aave_use_graph,
         max_tasks_num,
         legacy_messages_via_websockets,
+        data_migration_version,
+        use_custom_database,
+        user_data_dir,
+        perform_migrations_at_unlock,
 ):
     if not start_with_logged_in_user:
         return
@@ -147,11 +165,31 @@ def initialize_mock_rotkehlchen_instance(
     # does not run during tests
     size_patch = patch('rotkehlchen.rotkehlchen.ICONS_BATCH_SIZE', new=0)
     sleep_patch = patch('rotkehlchen.rotkehlchen.ICONS_QUERY_SLEEP', new=999999)
-    with settings_patch, eth_rpcconnect_patch, ksm_rpcconnect_patch, ksm_connect_on_startup_patch, size_patch, sleep_patch:  # noqa: E501
+
+    create_new = True
+    if use_custom_database is not None:
+        _use_prepared_db(user_data_dir, use_custom_database)
+        create_new = False
+
+    with ExitStack() as stack:
+        stack.enter_context(settings_patch)
+        stack.enter_context(eth_rpcconnect_patch)
+        stack.enter_context(ksm_rpcconnect_patch)
+        stack.enter_context(ksm_connect_on_startup_patch)
+        stack.enter_context(size_patch)
+        stack.enter_context(sleep_patch)
+        if perform_migrations_at_unlock is False:
+            migrations_patch = patch.object(
+                DataMigrationManager,
+                'maybe_migrate_data',
+                side_effect=lambda *args: None,
+            )
+            stack.enter_context(migrations_patch)
+
         rotki.unlock_user(
             user=username,
             password=db_password,
-            create_new=True,
+            create_new=create_new,
             sync_approval='no',
             premium_credentials=None,
         )
@@ -174,7 +212,7 @@ def initialize_mock_rotkehlchen_instance(
     # After unlocking when all objects are created we need to also include
     # customized fixtures that may have been set by the tests
     rotki.chain_manager.accounts = blockchain_accounts
-    add_settings_to_test_db(rotki.data.db, db_settings, ignored_assets)
+    add_settings_to_test_db(rotki.data.db, db_settings, ignored_assets, data_migration_version)
     maybe_include_etherscan_key(rotki.data.db, include_etherscan_key)
     maybe_include_cryptocompare_key(rotki.data.db, include_cryptocompare_key)
     add_blockchain_accounts_to_db(rotki.data.db, blockchain_accounts)
@@ -244,6 +282,10 @@ def fixture_rotkehlchen_api_server(
         aave_use_graph,
         max_tasks_num,
         legacy_messages_via_websockets,
+        data_migration_version,
+        use_custom_database,
+        user_data_dir,
+        perform_migrations_at_unlock,
 ):
     """A partially mocked rotkehlchen server instance"""
 
@@ -278,6 +320,10 @@ def fixture_rotkehlchen_api_server(
         aave_use_graph=aave_use_graph,
         max_tasks_num=max_tasks_num,
         legacy_messages_via_websockets=legacy_messages_via_websockets,
+        data_migration_version=data_migration_version,
+        use_custom_database=use_custom_database,
+        user_data_dir=user_data_dir,
+        perform_migrations_at_unlock=perform_migrations_at_unlock,
     )
     yield api_server
     api_server.stop()
@@ -309,6 +355,10 @@ def rotkehlchen_instance(
         aave_use_graph,
         max_tasks_num,
         legacy_messages_via_websockets,
+        data_migration_version,
+        use_custom_database,
+        user_data_dir,
+        perform_migrations_at_unlock,
 ):
     """A partially mocked rotkehlchen instance"""
 
@@ -337,6 +387,10 @@ def rotkehlchen_instance(
         aave_use_graph=aave_use_graph,
         max_tasks_num=max_tasks_num,
         legacy_messages_via_websockets=legacy_messages_via_websockets,
+        data_migration_version=data_migration_version,
+        use_custom_database=use_custom_database,
+        user_data_dir=user_data_dir,
+        perform_migrations_at_unlock=perform_migrations_at_unlock,
     )
     return uninitialized_rotkehlchen